% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mixpanelR.R
\docType{methods}
\name{export,MixPanel-method}
\alias{export,MixPanel-method}
\title{High-level API request for exporting events}
\usage{
\S4method{export}{MixPanel}(obj, event = NULL, params = NULL)
}
\arguments{
\item{obj}{MixPanel object}

\item{event}{The single event to be exported}

\item{params}{The list of parameters of API export request}
}
\value{
The data.frame containing the exported data
}
\description{
Produces an \code{export} request for the specified single event
and accounting for the specified parameters. Note that the parameters
\code{from_date} and \code{to_date} are required.
}
\examples{
\dontrun{
library(mixpanelR)

api <- MixPanel(
     api_key = "<API_KEY>"
     api_secret = "<API_SECRET>"
)

df <-  export(api,
     event = "project:successful_donation",
     params = list(from_date = "2015-01-01",
                     to_date = "2015-03-05")
)

summary(df)}
}

