% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mixpanelR.R
\name{export}
\alias{export}
\title{S4 generic dispatch method}
\usage{
export(obj, ...)
}
\arguments{
\item{obj}{Object}

\item{...}{arguments for the corresponding method}
}
\description{
S4 generic dispatch method
}
\seealso{
\code{\link{export,MixPanel-method}}
}

