% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mixpanelR.R
\docType{class}
\name{MixPanel-class}
\alias{MixPanel}
\alias{MixPanel-class}
\title{MixPanel Data Export API class}
\usage{
MixPanel(...)
}
\arguments{
\item{...}{Constructor parameters}
}
\description{
MixPanel Data Export API class
}
\section{Slots}{

\describe{
\item{\code{api_key}}{MixPanel API key}

\item{\code{api_secret}}{MixPanel API secret}

\item{\code{END_POINT}}{Default API endpoint (each request may specify a different endpoint if needed)}

\item{\code{VERSION}}{API version}
}}
\examples{
\dontrun{
library(mixpanelR)

api <- MixPanel(
     api_key = "<API_KEY>"
     api_secret = "<API_SECRET>"
)
}
}
\seealso{
\code{\link{request,MixPanel-method}, \link{export,MixPanel-method}}
}

