% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_export_api.R
\name{mixpanel_export_events}
\alias{mixpanel_export_events}
\title{Title}
\usage{
mixpanel_export_events(
  project_id,
  from_date,
  to_date,
  limit = NULL,
  event = NULL,
  where = NULL,
  username,
  secret,
  eu_region = FALSE
)
}
\arguments{
\item{project_id}{required, an integer representing the id of the project}

\item{from_date}{required string, The date in yyyy-mm-dd format to query to. This date is inclusive.}

\item{to_date}{required string, The date in yyyy-mm-dd format from which to begin querying from. This date is inclusive.}

\item{limit}{optional integer, The max number of events to be returned.}

\item{event}{optional vector, The event or events that you wish to get data for}

\item{where}{optional string, An expression to filter events by. More info on expression sequence structure can be found [here](https://developer.mixpanel.com/reference-link/segmentation-expressions)}

\item{username}{required, service account username see https://developer.mixpanel.com/reference/service-accounts}

\item{secret}{required, service account secret see https://developer.mixpanel.com/reference/service-accounts}

\item{eu_region}{optional, set to TRUE if you are enrolled in the EU server residency}
}
\value{
returns a tibble
}
\description{
Title
}
