% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_api_insights.R
\name{mixpanel_insights}
\alias{mixpanel_insights}
\title{Query saved insights report}
\usage{
mixpanel_insights(
  project_id,
  bookmark_id,
  workspace_id = NULL,
  username,
  secret,
  eu_region = FALSE
)
}
\arguments{
\item{project_id}{required, an integer representing the id of the project}

\item{bookmark_id}{required, bookmark_id of your Insights report can be found from the url: https://mixpanel.com/report/1/insights#report/<YOUR_BOOKMARK_ID>/example-report}

\item{workspace_id}{optional, the id of the workspace if applicable.}

\item{username}{required, service account username https://developer.mixpanel.com/reference/service-accounts}

\item{secret}{required, service account secret https://developer.mixpanel.com/reference/service-accounts}

\item{eu_region}{optional, set to TRUE if you are enrolled in the EU server residency}
}
\value{
returns a tibble with the data from the insights report
}
\description{
Return data from a saved insights report in Mixpanel
}
\examples{
\dontrun{
 mixpanel_insights()
 }
}
