% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mixpanelR.R
\docType{methods}
\name{request,MixPanel-method}
\alias{request,MixPanel-method}
\title{MixPanel Data Export API request method}
\usage{
\S4method{request}{MixPanel}(obj, methods = NULL, params = NULL,
  format = "json", endpoint = NULL, presigned = FALSE)
}
\arguments{
\item{obj}{MixPanel object}

\item{methods}{The list of methods for the API call}

\item{params}{The list of parameters of the API call}

\item{format}{Response format (default: json)}

\item{endpoint}{URL of custom API endpoint}

\item{presigned}{Specifies whether the request is already signed (default: FALSE)}
}
\value{
The raw text of the response in the requested format (default: JSON)
}
\description{
MixPanel Data Export API request method
}
\examples{
\dontrun{
library(mixpanelR)
library(rjson)

api <- MixPanel(
     api_key = "<API_KEY>"
     api_secret = "<API_SECRET>"
)

raw_response <- request(api,
     methods = list("events"),
     params = list(
         event = list("project:successful_donation","project:loaded"),
         unit = "hour",
         interval = 24,
         type = "general"
     )
)

data <- fromJSON(raw_response)}
}

