% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mixpanelR.R
\name{request}
\alias{request}
\title{S4 generic dispatch method}
\usage{
request(obj, ...)
}
\arguments{
\item{obj}{Object}

\item{...}{arguments for the corresponding method}
}
\description{
S4 generic dispatch method
}
\seealso{
\code{\link{request,MixPanel-method}}
}

