% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights.R
\name{mixpanel_insights}
\alias{mixpanel_insights}
\title{Title}
\usage{
mixpanel_insights(
  project_id,
  bookmark_id,
  workspace_id = NULL,
  username,
  secret,
  eu_region = FALSE
)
}
\arguments{
\item{eu_region}{}
}
\value{

}
\description{
Title
}
